Routing Requests, Parse Requests, Redirecting request into a new file.
wrote code to route a request sent into the server through the browser in Node.js
wrote code to redirect the message and save message in a new file.
    also parses the message the user input and saves that data in a new file.
utilized on(), push(), concat() and toString() to gather the data that was input and concatenate it into a string.
parsed through a request body and concat() the message into a separate file.