Self Assignment 1
Created my own server with node.js using port 3000, created with two routes "/" and "/users".  returned a greetings on "/" and a dummy list on "/users".  Then added a form with a "username" <input> to the "/" page and submit a "POST" request to "/create-user" upon clicking the submit button.  Finally, I added the data coming in from "/create-user", parsed the data and logged it to the console.