two quick bug fixes
fixed a bug that ever,y once in a while when you log in, the redirect to the webpage would happen before postLogin would be done writing to the database and the view would not show some needed items on the navigation.

also fixed edit product in the details page when not logged inl.